Fix ModuleArchiveContainer tests and split retryImport into its own file
