+++ website/src/utils/__mocks__/error.ts
@@ -1,3 +1,2 @@
 export const captureException = jest.fn();
 export const getScriptErrorHandler = jest.fn().mockReturnValue(() => jest.fn());
-export const retryImport = jest.fn().mockResolvedValue(undefined);
