+++ website/src/views/contribute/ContributeContainer/index.tsx
@@ -2,7 +2,7 @@ import Loadable, { LoadingComponentProps } from 'react-loadable';
 
 import LoadingSpinner from 'views/components/LoadingSpinner';
 import ApiError from 'views/errors/ApiError';
-import { retryImport } from 'utils/error';
+import retryImport from 'utils/retryImport';
 
 const AsyncContributeContainer = Loadable({
   loader: () =>
